♻️ refactor(tests): mirror src layout + quality
Move every test module into the directory that mirrors the source it
exercises, the way tests/query/css already mirrored the C selector
engine. A test's path now names what it covers, and code and test move
together.

The 41-file tests/features catch-all splits by public namespace: the
sanitizer, linkifier, link detector and minifier tests to tests/clean;
dates, microdata, RDFa, Open Graph, Dublin Core, boilerplate, links and
URL cleaning to tests/extract; the URL primitives to tests/url; the
builder to tests/build. The stragglers find their mirror too: the
css_to_xpath suite backs convert.py so it moves to tests/convert,
test_vec_overflow mirrors _c/core/vec.h under tests/core, the encoding
detector API tests join tests/detect, and table treebuilding joins
tests/dom.

Quality fixes ride along: the duplicated microdata and Open Graph suites
fold into one file each (RDFa and Dublin Core split into their own
mirror files); the encoding-detection coverage drivers pin the exact
label each byte input yields rather than only that some label came back;
and the first fixture becomes first_of_type to end its clash with the
local first variables.

pytest's default norecursedirs skips any directory named build, so
tests/build needs an explicit override to be collected.

closes #33
Part of #478
