+++ tests/clean/data/dompurify_expect.mjs
