+++ tests/clean/test_detector_c_api.py
