+++ tests/clean/test_detector_find.py
