+++ tests/clean/test_detector_presence.py
