+++ tests/clean/test_detector_schemes.py
