+++ tests/clean/test_detector_span.py
