+++ tests/clean/test_minify.py
