+++ tests/clean/test_sanitizer_namespace.py
