+++ tests/core/test_vec_overflow.py
