+++ tests/detect/test_detect_api.py
