+++ tests/detect/test_detect_detector.py
