+++ tests/detect/test_detect_options.py
