+++ tests/extract/test_annotation_surface.py
