+++ tests/extract/test_annotation_tags.py
