+++ tests/extract/test_extract_opengraph.py
@@ -1,9 +1,95 @@
+"""opengraph: extract.opengraph / Node.opengraph(), the OpenGraph mapping, is_valid and URL resolution.
+
+The function entry point and the DOM method feed the same reader; the function cases pin the mapping
+protocol (getitem, repr, is_valid) while the parametrized method cases pin the tag-selection and
+base-URL rules.
+"""
+
 from __future__ import annotations
 
 import pytest
 
 from turbohtml import parse
-from turbohtml._structured_data import OpenGraph
+from turbohtml.extract import OpenGraph, opengraph
+
+_FULL = (
+    '<meta property="og:title" content="The Rock">'
+    '<meta property="og:type" content="video.movie">'
+    '<meta property="og:image" content="http://x/rock.jpg">'
+    '<meta property="og:url" content="http://x/tt0117500/">'
+)
+
+
+def test_opengraph_strips_prefix() -> None:
+    assert opengraph('<meta property="og:title" content="Hello">') == {"title": "Hello"}
+
+
+def test_opengraph_drops_twitter_tags() -> None:
+    html = '<meta property="og:title" content="Hi"><meta name="twitter:card" content="summary">'
+    assert dict(opengraph(html)) == {"title": "Hi"}
+
+
+def test_opengraph_keeps_structured_subkeys() -> None:
+    html = '<meta property="og:image" content="http://x/a.png"><meta property="og:image:width" content="400">'
+    assert opengraph(html) == {"image": "http://x/a.png", "image:width": "400"}
+
+
+def test_opengraph_empty_without_tags() -> None:
+    assert dict(opengraph("<p>plain</p>")) == {}
+
+
+def test_getitem_reads_stripped_key() -> None:
+    assert opengraph('<meta property="og:title" content="Hello">')["title"] == "Hello"
+
+
+def test_missing_key_raises() -> None:
+    with pytest.raises(KeyError):
+        _ = opengraph('<meta property="og:title" content="Hello">')["type"]
+
+
+def test_membership_and_get() -> None:
+    og = opengraph('<meta property="og:title" content="Hello">')
+    assert "title" in og
+    assert og.get("missing") is None
+
+
+def test_repr_wraps_properties() -> None:
+    assert repr(opengraph('<meta property="og:title" content="Hi">')) == "OpenGraph({'title': 'Hi'})"
+
+
+def test_is_valid_with_all_four_required() -> None:
+    assert opengraph(_FULL).is_valid() is True
+
+
+@pytest.mark.parametrize(
+    "html",
+    [
+        pytest.param('<meta property="og:title" content="only">', id="one-property"),
+        pytest.param("<p>no meta</p>", id="no-properties"),
+    ],
+)
+def test_is_valid_false_when_required_missing(html: str) -> None:
+    assert opengraph(html).is_valid() is False
+
+
+def test_is_valid_false_when_required_empty() -> None:
+    html = _FULL + '<meta property="og:url" content="">'
+    assert opengraph(html).is_valid() is False
+
+
+def test_direct_construction() -> None:
+    og = OpenGraph({"title": "Hi", "type": "article", "image": "i", "url": "u"})
+    assert og.is_valid() is True
+    assert len(og) == 4
+
+
+def test_opengraph_function_resolves_url_against_base() -> None:
+    html = '<meta property="og:image" content="/rock.jpg"><meta property="og:title" content="R">'
+    assert opengraph(html, "http://x.com/dir/") == {"image": "http://x.com/rock.jpg", "title": "R"}
+
+
+def test_opengraph_function_base_url_omitted_is_verbatim() -> None:
+    assert opengraph('<meta property="og:image" content="/rock.jpg">') == {"image": "/rock.jpg"}
 
 
 @pytest.mark.parametrize(
@@ -61,7 +147,7 @@ from turbohtml._structured_data import OpenGraph
         pytest.param("<p>not a meta</p>", {}, id="no-meta"),
     ],
 )
-def test_opengraph(html: str, expected: dict[str, str]) -> None:
+def test_opengraph_document_method(html: str, expected: dict[str, str]) -> None:
     assert parse(html).opengraph() == expected
 
 
@@ -172,33 +258,3 @@ def test_opengraph_base_url_wrong_type_raises() -> None:
 def test_opengraph_rejects_unexpected_argument() -> None:
     with pytest.raises(TypeError):
         parse("<p>x</p>").opengraph(nope=1)  # ty: ignore[unknown-argument]  # exercises the argument-parse failure
-
-
-@pytest.mark.parametrize(
-    ("html", "expected"),
-    [
-        pytest.param(
-            '<meta name="dc.title" content="Hi"><meta name="dc.creator" content="Ada">',
-            {"dc.title": "Hi", "dc.creator": "Ada"},
-            id="dc-names",
-        ),
-        pytest.param('<meta name="dcterms.created" content="2020">', {"dcterms.created": "2020"}, id="dcterms-name"),
-        pytest.param('<meta name="DC.Title" content="Hi">', {"dc.title": "Hi"}, id="name-lower-cased"),
-        pytest.param(
-            '<meta name="dc.title" content="first"><meta name="dc.title" content="second">',
-            {"dc.title": "second"},
-            id="last-occurrence-wins",
-        ),
-        pytest.param('<meta name="dc.title">', {"dc.title": ""}, id="missing-content-empty-string"),
-        pytest.param('<meta name="dc.title" content>', {"dc.title": ""}, id="valueless-content-empty-string"),
-        pytest.param('<meta name="keywords" content="x">', {}, id="non-dc-name-ignored"),
-        pytest.param('<meta name="dc" content="x">', {}, id="name-shorter-than-prefix-ignored"),
-        pytest.param('<meta name="dcx.title" content="x">', {}, id="near-miss-prefix-ignored"),
-        pytest.param('<meta property="dc.title" content="x">', {}, id="property-not-name-ignored"),
-        pytest.param('<meta name content="x">', {}, id="valueless-name-ignored"),
-        pytest.param('<meta content="x">', {}, id="no-name-ignored"),
-        pytest.param("<p>not a meta</p>", {}, id="no-meta"),
-    ],
-)
-def test_dublin_core(html: str, expected: dict[str, str]) -> None:
-    assert parse(html).dublin_core() == expected
