+++ tests/extract/test_links_resolve.py
