+++ tests/extract/test_structured_data_combined.py
